enhance: download and compute the hash with one read
Before the change, the file is read twice:
* once for the download
* once when checking the MD5

With this change, the MD5 is computed at download time.
This can be useful if the files are big as reads can be expensive.

Also, the temporary file’s name is generated using
tempfile.NamedTemporaryFile. This can help with parallelisation (as it
avoids filename collisions) and it also leverages the `/tmp` directory
that’s automatically cleaned up on reboots and doesn’t clutter other
folders.